4.6 Добавит структуру данных фильтров.
