+++ src/mock/event-filter.js
@@ -1,0 +1,7 @@
+import {eventFilter} from '../utils/event-filter.js';
+
+export const generateEventFilter = () => Object.entries(eventFilter).map(
+  ([filterName]) => ({
+    name: filterName,
+  }),
+);
