+++ src/utils/event.js
@@ -5,3 +5,4 @@ export const humanizeDateToTimeDateMini = (data) => dayjs(data).format('YYYY-MM-
 export const humanizeDate = (date) => dayjs(date).format('DD/MM/YY HH:mm');
 export const humanizeDateToTime = (date) => dayjs(date).format('HH:mm');
 export const humanizeDateToDayMonth = (data) => dayjs(data).format('MMM DD');
+export const isEventActive = (eventDate) => eventDate && dayjs().isBefore(eventDate, 'D');
